separate tests using test helper
