+++ core/lib/zksync_core/src/eth_sender/mod.rs
@@ -5,6 +5,7 @@ mod eth_tx_aggregator;
 mod eth_tx_manager;
 mod metrics;
 mod publish_criterion;
+mod tests_helper;
 mod zksync_functions;
 
 #[cfg(test)]
